+++ resources/js/Components/DatePicker/index.ts
@@ -1,0 +1,6 @@
+export { default as AppDateInput } from "./AppDateInput";
+export {
+    default as AppDateRangeInput,
+    DEFAULT_PERIOD_PRESETS,
+} from "./AppDateRangeInput";
+export type {} from "./AppDateRangeInput";